requirements.txt
# Core data & ML
yfinance==1.2.0
pandas==2.3.3
numpy==1.26.4
matplotlib==3.10.8
scikit-learn==1.8.0
PyYAML==6.0.3
joblib==1.5.3

# Gradient boosting
xgboost==3.2.0
lightgbm==4.6.0

# Hyperparameter optimization
optuna==4.8.0

# Sentiment & NLP
textblob==0.19.0
snownlp==0.12.3
pytrends==4.9.2

# Time series features
tsfresh==0.21.1
pyts==0.13.0

# Backtesting
vectorbt==0.28.4

# Utilities
requests==2.32.5
tenacity==9.1.4
exchange_calendars==4.13.2
pyarrow==23.0.1

# Security scanning (for CI)
pip-audit==2.10.0
